third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/xsl-img-blocked-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
CONSOLE ERROR: Loading the image 'http://127.0.0.1:8000/security/resources/abe.png' violates the following Content Security Policy directive: "img-src 'none'". The action has been blocked.
Here is an image: 
